Sandbox class only deletes .git if it exists

+++ GRID_PiCaS_Launcher/sandbox.py
@@ -65,7 +65,7 @@ class Sandbox(object):
         if repo_commit:
             checkout = subprocess.Popen(['git', 'checkout', repo_commit])
             checkout.wait()
-        if remove_gitdir:
+        if remove_gitdir and os.path.isdir('/.git'):
             shutil.rmtree('.git/')
         if not checkout_dir:
             files = os.listdir(checkout_dir_path)
